refactor(order): move icons folder
