+++ package.json
@@ -47,15 +47,9 @@
   },
   "homepage": "https://corona-mood.firebaseapp.com/",
   "dependencies": {
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "@testing-library/user-event": "^12.6.2",
     "babel-plugin-formatjs": "^9.0.1",
     "firebase": "^8.2.4",
-    "formatjs": "^0.1.1",
     "moment": "^2.29.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
@@ -68,6 +62,10 @@
   },
   "devDependencies": {
     "@formatjs/cli": "^3.0.1",
+    "@testing-library/jest-dom": "^5.11.9",
+    "@testing-library/react": "^11.2.3",
+    "@testing-library/user-event": "^12.6.2",
+    "formatjs": "^0.1.1",
     "jest-styled-components": "^7.0.3",
     "match-media-mock": "^0.1.1",
     "react-intl-cra": "^0.3.4",
