+++ src/components/Styles/common.styles.js
@@ -1,5 +1,6 @@
-import { createGlobalStyle } from 'styled-components';
+import styled, { css, createGlobalStyle } from 'styled-components';
 
+import { ReactComponent as LogoSVG }  from '../../assets/images/sentirama-logo.svg';
 
 const bodyOverflowHidden = createGlobalStyle`
     body {
@@ -13,7 +14,15 @@ const bodyOverflowVisible = createGlobalStyle`
     }
 `;
 
+const SentiramaLogo = styled(LogoSVG)`
+    ${({ theme }) => css`
+        fill: ${theme.colorMainLogo || theme.neutral900};
+        height: ${theme.logoDefaultHeight};
+    `}
+`;
+
 export {
     bodyOverflowHidden,
     bodyOverflowVisible,
+    SentiramaLogo,
 };
